requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9

# AI/ML
google-generativeai==0.7.1

# Async/HTTP
anyio==4.3.0
httpx==0.27.0

# Data validation
pydantic==2.7.0

# Production
python-multipart==0.0.9
